requirements.txt
python-telegram-bot==13.18
requests
beautifulsoup4